feat(engine): promote world positions to double precision (#170) (#177)
Widen EntityTransform::pos[3], MsgEntityEntry::pos[3] (56→68 bytes),
EntityRenderEntry::position, and CameraView::worldOrigin from float/
glm::vec3 to double/glm::dvec3 throughout the sim→net→render pipeline.
GPU math stays float32 via camera-relative rendering — relPos is computed
as glm::vec3(dvec3_worldPos - dvec3_worldOrigin) before upload.

+++ engine/entity/EntityState.h
@@ -10,7 +10,7 @@
 namespace fl {
 
 struct EntityTransform {
-    float pos[3]{};            // world position (m)
+    double pos[3]{};           // world position (m) — double for planet-scale precision
     float vel[3]{};            // world velocity (m/s)
     float quat[4]{0, 0, 0, 1}; // orientation quaternion [x,y,z,w]
 };
